requirements.txt
pyinstaller==5.13.2
ffmpeg-python==0.2.0
